modify periodic distribution framework
